Various compiler warnings.

+++ libls/print.c
@@ -137,7 +137,7 @@ printlong (DISPLAY *dp)
       if (f_flags)
 	printf ("%-*s ", dp->s_flags, np->flags);
       if (S_ISCHR (sp->st_mode) || S_ISBLK (sp->st_mode))
-	printf ("%3d, %3d ", major (sp->st_rdev), minor (sp->st_rdev));
+	printf ("%3d, %3d ", (int) major (sp->st_rdev), (int) minor (sp->st_rdev));
       else if (dp->bcfile)
 	printf ("%*s%*llu ",
 		8 - dp->s_size, "", dp->s_size, (long long) sp->st_size);
